requirements.txt
prompt-toolkit==3.0.2
python-slugify==4.0.0

